great

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -22,5 +22,6 @@ int _atoi(char *s)
 			break;
 		i++;
 	}
-	return (num * sign);
+	
+	return (int)(num * sign);
 }
